requirements.txt
boto3>=1.34.0
langchain>=0.1.20
langgraph>=0.0.35
langchain-aws>=0.1.6
langchain-core>=0.1.45
langchain-community>=0.0.32
langchainhub>=0.1.15
typing-extensions>=4.7.0
pypdf>=3.17.4
tiktoken>=0.6.0
python-dotenv>=1.0.1
streamlit>=1.35.0